Selection index for empty content, subview layout.
Instead of updating the subview layout in -[MBCoverFlowView
viewWillMoveToSuperview], we now dow this in -setFrame:. The subview
layout doesn't depend on the cover flow view's superview, but it does
depend on the frame size.

Indicate an empty selection (i. e. no content) with a selectionIndex
of NSNotFound.

Use an animation duration of 0.0 when changing the selection index
while an MBCoverFlowView is not in a window (which means it isn't
visible).